Rename class ColumnSchemaBuilder
