--- src/ColumnSchemaBuilder.php
@@ -1,31 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Yiisoft\Db\Sqlite;
-
-use Yiisoft\Db\Schema\AbstractColumnSchemaBuilder;
-
-final class ColumnSchemaBuilder extends AbstractColumnSchemaBuilder
-{
-    /**
-     * Builds the unsigned string for column. Defaults to unsupported.
-     *
-     * @return string a string containing UNSIGNED keyword.
-     */
-    protected function buildUnsignedString(): string
-    {
-        return $this->isUnsigned() ? ' UNSIGNED' : '';
-    }
-
-    public function asString(): string
-    {
-        $format = match ($this->getTypeCategory()) {
-            self::CATEGORY_PK => '{type}{check}{append}',
-            self::CATEGORY_NUMERIC => '{type}{length}{unsigned}{notnull}{unique}{check}{default}{append}',
-            default => '{type}{length}{notnull}{unique}{check}{default}{append}',
-        };
-
-        return $this->buildCompleteString($format);
-    }
-}
